Better support for large SHOWPLANs and code cleanup
Large SHOWPLANS remain to be brutal. Some were so big you'd be looking at
a blank page until you realized you need to scroll to the right. This at
least gets the first node centered in the display via a javascript call.
Not the best solution, but sufficient for now.

Also did some widespread cleanup of the SCSS files to help with their
read-ability and also made the node colors an explicit color rather than
transparent.