Used subtree cost for operator cost
Using the EstimatedIo and EstimatedCpu along with rebinds and rewinds proved to be close but not spot on. Resulted in queries with 110% usage and other oddities. Seems the only way to get it to match SMSS is to use estimated subtree cost - sum(children's sub tree cost) to get the proper value. This still has an odditity where some queries drop below zero for the cost. Seems SMSS rounds up to zero in those cases.